# internal/server/handler/organization.go
package handler

import (
	"net/http"

	"git.ocn.com.vn/ocn/common/httpbase"
	"github.com/quanNaNbk03/monitor-demo/internal/server/service"
	org "github.com/quanNaNbk03/monitor-demo/pkg/api/organization"

	"github.com/gin-gonic/gin"
)

// OrgHandler implements the org.ServerInterface generated by oapi-codegen
type OrgHandler struct {
	orgService service.OrgService
}

// NewOrgHandler creates a new org handler
func NewOrgHandler(orgService service.OrgService) *OrgHandler {
	return &OrgHandler{
		orgService: orgService,
	}
}

// Ensure OrgHandler implements org.ServerInterface
var _ org.ServerInterface = (*OrgHandler)(nil)

func (h *OrgHandler) Create(c *gin.Context) {
	req := new(org.CreateOrgRequest)
	if err := httpbase.BindInput(c, req); err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}

	if err := h.orgService.Create(c.Request.Context(), req); err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}

	httpbase.ReturnSuccessResponse(c, http.StatusOK, nil)
}

func (h *OrgHandler) Get(c *gin.Context, id uint64) {
	res, err := h.orgService.Get(c.Request.Context(), id)
	if err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}
	httpbase.ReturnSuccessResponse(c, http.StatusOK, res)
}

func (h *OrgHandler) Delete(c *gin.Context, id uint64) {
	if err := h.orgService.Delete(c.Request.Context(), id); err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}
	httpbase.ReturnSuccessResponse(c, http.StatusOK, nil)
}

func (h *OrgHandler) Update(c *gin.Context, id uint64) {
	in := new(org.UpdateOrgRequest)
	if err := httpbase.BindInput(c, in); err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}
	if err := h.orgService.Update(c.Request.Context(), id, in); err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}
	httpbase.ReturnSuccessResponse(c, http.StatusOK, nil)
}

func (h *OrgHandler) ListOrgs(c *gin.Context) {
	in := new(org.ListOrgsJSONBody)
	if ierr := httpbase.BindInput(c, in); ierr != nil {
		httpbase.ReturnErrorResponse(c, ierr)
		return
	}
	res, err := h.orgService.List(c.Request.Context(), in)
	if err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}
	httpbase.ReturnSuccessResponse(c, http.StatusOK, res)
}