# internal/server/handler/zone.go
package handler

import (
	"net/http"

	"git.ocn.com.vn/ocn/common/httpbase"
	"github.com/quanNaNbk03/monitor-demo/internal/server/dto"
	"github.com/quanNaNbk03/monitor-demo/internal/server/service"
	"github.com/quanNaNbk03/monitor-demo/pkg/api/zone"

	"github.com/gin-gonic/gin"
)

// ZoneHandler implements the health.ServerInterface generated by oapi-codegen
type ZoneHandler struct {
	zoneService service.ZoneService
}

// NewZoneHandler creates a new health handler
func NewZoneHandler(zoneService service.ZoneService) *ZoneHandler {
	return &ZoneHandler{
		zoneService: zoneService,
	}
}

// Ensure ZoneHandler implements zone.ServerInterface
var _ zone.ServerInterface = (*ZoneHandler)(nil)

func (h *ZoneHandler) Get(c *gin.Context, id uint64) {
	res, err := h.zoneService.Get(c.Request.Context(), id)
	if err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}

	httpbase.ReturnSuccessResponse(c, http.StatusOK, res)
}

func (h *ZoneHandler) Create(c *gin.Context) {
	req := new(zone.CreateZoneRequest)
	if err := httpbase.BindInput(c, req); err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}

	if err := h.zoneService.Create(c.Request.Context(), req); err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}

	httpbase.ReturnSuccessResponse(c, http.StatusOK, nil)
}

func (h *ZoneHandler) Update(c *gin.Context, id uint64) {
	req := new(zone.UpdateZoneRequest)
	if err := httpbase.BindInput(c, req); err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}

	if err := h.zoneService.Update(c.Request.Context(), id, &dto.UpdateZoneRequest{
		Type:    req.Type,
		SoaType: req.SoaType,
	}); err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}
	httpbase.ReturnSuccessResponse(c, http.StatusOK, nil)
}

func (h *ZoneHandler) ListZones(c *gin.Context) {
	in := new(dto.ListZonesJSONBody)
	if err := httpbase.BindInput(c, in); err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}
	res, err := h.zoneService.List(c.Request.Context(), in)
	if err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}
	httpbase.ReturnSuccessResponse(c, http.StatusOK, res)
}

func (h *ZoneHandler) Delete(c *gin.Context, id uint64) {
	if err := h.zoneService.Delete(c.Request.Context(), id); err != nil {
		httpbase.ReturnErrorResponse(c, err)
		return
	}
	httpbase.ReturnSuccessResponse(c, http.StatusOK, nil)
}