# pkg/api/organization/organization.gen.go
// Package org provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package org

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	"github.com/quanNaNbk03/monitor-demo/internal/server/dto"
)

// CreateOrgRequest defines model for CreateOrgRequest.
type CreateOrgRequest struct {
	Contact     string `json:"contact" validate:"required,max=255"`
	Description string `json:"description,omitempty"`
	Email       string `json:"email" validate:"required,max=255"`
	Name        string `json:"name" validate:"required,max=255"`
}

// Error defines model for Error.
type Error struct {
	Error struct {
		Code    string `json:"code,omitempty"`
		Detail  string `json:"detail,omitempty"`
		Message string `json:"message,omitempty"`
		Name    string `json:"name,omitempty"`
	} `json:"error,omitempty"`
}

// ListOrganizationResponse defines model for ListOrganizationResponse.
type ListOrganizationResponse struct {
	Organizations []Organization `json:"organizations,omitempty"`
	Paging        Paging         `json:"paging,omitempty"`
}

// Organization defines model for Organization.
type Organization = dto.Organization

// Paging defines model for Paging.
type Paging = dto.Paging

// PagingRequest defines model for PagingRequest.
type PagingRequest struct {
	Limit int `json:"limit,omitempty"`
	Page  int `json:"page,omitempty"`
}

// SearchRequest defines model for SearchRequest.
type SearchRequest struct {
	Field string      `json:"field,omitempty"`
	Op    string      `json:"op,omitempty"`
	Value interface{} `json:"value,omitempty"`
}

// UpdateOrgRequest defines model for UpdateOrgRequest.
type UpdateOrgRequest struct {
	Contact     *string `json:"contact,omitempty" validate:"required,max=255"`
	Description *string `json:"description,omitempty"`
	Email       *string `json:"email,omitempty" validate:"required,max=255"`
}

// ErrorResponse400 defines model for ErrorResponse400.
type ErrorResponse400 = Error

// ErrorResponse401 defines model for ErrorResponse401.
type ErrorResponse401 = Error

// ErrorResponse403 defines model for ErrorResponse403.
type ErrorResponse403 = Error

// ErrorResponse404 defines model for ErrorResponse404.
type ErrorResponse404 = Error

// ErrorResponse500 defines model for ErrorResponse500.
type ErrorResponse500 = Error

// CreateJSONBody defines parameters for Create.
type CreateJSONBody = []CreateOrgRequest

// ListOrgsJSONBody defines parameters for ListOrgs.
type ListOrgsJSONBody struct {
	Paging   PagingRequest   `json:"paging,omitempty"`
	Searches []SearchRequest `json:"searches,omitempty"`
}

// CreateJSONRequestBody defines body for Create for application/json ContentType.
type CreateJSONRequestBody = CreateJSONBody

// ListOrgsJSONRequestBody defines body for ListOrgs for application/json ContentType.
type ListOrgsJSONRequestBody ListOrgsJSONBody

// UpdateJSONRequestBody defines body for Update for application/json ContentType.
type UpdateJSONRequestBody = UpdateOrgRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create Create new Oraganization
	// (POST /api/v1/organizations)
	Create(c *gin.Context)
	// ListOrgs List all Organizations
	// (POST /api/v1/organizations/search)
	ListOrgs(c *gin.Context)
	// Delete delete a Oraganization
	// (DELETE /api/v1/organizations/{id})
	Delete(c *gin.Context, id uint64)
	// Get get a Oraganization
	// (GET /api/v1/organizations/{id})
	Get(c *gin.Context, id uint64)
	// Update Update a Organization
	// (PUT /api/v1/organizations/{id})
	Update(c *gin.Context, id uint64)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// Create operation middleware
func (siw *ServerInterfaceWrapper) Create(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Create(c)
}

// ListOrgs operation middleware
func (siw *ServerInterfaceWrapper) ListOrgs(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListOrgs(c)
}

// Delete operation middleware
func (siw *ServerInterfaceWrapper) Delete(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id uint64

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "uint64", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Delete(c, id)
}

// Get operation middleware
func (siw *ServerInterfaceWrapper) Get(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id uint64

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "uint64", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Get(c, id)
}

// Update operation middleware
func (siw *ServerInterfaceWrapper) Update(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id uint64

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "uint64", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Update(c, id)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/api/v1/organizations", wrapper.Create)
	router.POST(options.BaseURL+"/api/v1/organizations/search", wrapper.ListOrgs)
	router.DELETE(options.BaseURL+"/api/v1/organizations/:id", wrapper.Delete)
	router.GET(options.BaseURL+"/api/v1/organizations/:id", wrapper.Get)
	router.PUT(options.BaseURL+"/api/v1/organizations/:id", wrapper.Update)
}